requirements.txt
redis==5.0.1
requests==2.31.0
paho-mqtt==1.6.1
